# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.FriendVo:
    properties:
      desc:
        description: 简介
        type: string
      device_info:
        description: 客户端设备信息
        type: string
      nickname:
        description: 昵称
        type: string
      online_status:
        allOf:
        - $ref: '#/definitions/model.OnlineStatus'
        description: 用户在线状态（如在线、离线、忙碌等）
      user_id:
        description: 用户ID
        type: integer
    type: object
  model.GroupAnnouncementCreateRequest:
    properties:
      content:
        type: string
      publisher:
        type: integer
    type: object
  model.GroupAnnouncementUpdateRequest:
    properties:
      announcement_id:
        description: 公告ID
        type: integer
      content:
        description: 公告内容
        type: string
    type: object
  model.GroupCreateRequest:
    properties:
      max_num:
        description: 群最大人数
        maximum: 2000
        minimum: 1
        type: integer
      member_list:
        description: 初始群成员
        items:
          type: integer
        type: array
      name:
        description: 群名称
        maxLength: 50
        type: string
      user_id:
        description: 创建者ID
        type: integer
    required:
    - name
    - user_id
    type: object
  model.GroupMemberMuteRequest:
    properties:
      duration:
        description: 禁言时长（秒）
        type: integer
      member_id:
        description: 要禁言的成员ID
        type: integer
    required:
    - duration
    - member_id
    type: object
  model.GroupMuteRequest:
    properties:
      group_id:
        type: integer
      mute_end:
        type: integer
    required:
    - group_id
    - mute_end
    type: object
  model.GroupSearchRequest:
    properties:
      code:
        description: 群号，精准查询
        type: string
      name:
        description: 群名称，模糊查询
        type: string
      page:
        type: integer
      pageSize:
        type: integer
      user_id:
        description: 用户ID(我加入的群)
        type: integer
    type: object
  model.GroupTransferRequest:
    properties:
      group_id:
        description: 群ID
        type: integer
      new_owner_id:
        description: 新群主的用户ID
        type: integer
    required:
    - group_id
    - new_owner_id
    type: object
  model.GroupUpdateRequest:
    properties:
      avatar:
        type: string
      desc:
        type: string
      group_id:
        type: integer
      max_num:
        type: integer
      name:
        type: string
    required:
    - group_id
    type: object
  model.GroupVo:
    properties:
      friend_vo_list:
        items:
          $ref: '#/definitions/model.FriendVo'
        type: array
      group_id:
        type: integer
      name:
        type: string
    type: object
  model.KickMemberRequest:
    properties:
      member_id:
        description: 被踢的成员ID
        type: integer
    required:
    - member_id
    type: object
...
  model.MemberVo:
    properties:
      avatar:
        type: string
      group_id:
        type: integer
      mute_end:
        type: string
      nickname:
        type: string
      online_status:
        $ref: '#/definitions/model.OnlineStatus'
      role:
        $ref: '#/definitions/model.Role'
      user_id:
        type: integer
    type: object
...
  model.OnlineStatus:
    enum:
    - 0
    - 1
    - 2
    - 3
    type: integer
    x-enum-comments:
      Away: 3 离开
      Busy: 2 忙碌
      Offline: 0 离线
      Online: 1 在线
    x-enum-descriptions:
    - 0 离线
    - 1 在线
    - 2 忙碌
    - 3 离开
    x-enum-varnames:
    - Offline
    - Online
    - Busy
    - Away
...
  model.Response:
    description: 通用响应格式，用于包装所有 API 响应
    properties:
      code:
        description: 状态码
        type: integer
      data:
        description: 数据部分
      message:
        description: 提示信息
        type: string
    type: object
  model.Role:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-comments:
      Admin: 1 管理员
      Member: 0 成员
      Owner: 2 群主
    x-enum-descriptions:
    - 0 成员
    - 1 管理员
    - 2 群主
    x-enum-varnames:
    - Member
    - Admin
    - Owner
  model.SetAdminRequest:
    properties:
      member_id:
        type: integer
    required:
    - member_id
    type: object
...
  model.UnmuteMemberRequest:
    properties:
      member_id:
        type: integer
    required:
    - member_id
    type: object
  model.UnsetAdminRequest:
    properties:
      member_id:
        description: 被取消管理员权限的用户 ID
        type: integer
    required:
    - member_id
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  title: Swagger Example API
  version: "1.0"
paths:
...
  /group/{group_id}/announcement:
    get:
      description: 获取指定群组的公告内容
      parameters:
      - description: 群组ID
        in: path
        name: group_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: 查询群组公告
      tags:
      - Group
  /group/{group_id}/announcement/create:
    post:
      consumes:
      - application/json
      description: 创建一个新的群组公告
      parameters:
      - description: 群组ID
        in: path
        name: group_id
        required: true
        type: integer
      - description: 公告内容
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.GroupAnnouncementCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: 创建群组公告
      tags:
      - Group
  /group/{group_id}/announcement/delete:
    get:
      consumes:
      - application/json
      description: 删除指定群组的公告
      parameters:
      - description: 群组ID
        in: path
        name: group_id
        required: true
        type: integer
      - description: 公告ID
        in: query
        name: announcement_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: 删除群组公告
      tags:
      - Group
  /group/{group_id}/announcement/update:
    post:
      consumes:
      - application/json
      description: 更新指定的群组公告
      parameters:
      - description: 群组ID
        in: path
        name: group_id
        required: true
        type: integer
      - description: 公告更新内容
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.GroupAnnouncementUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: 更新群组公告
      tags:
      - Group
  /group/{group_id}/announcement_list:
    get:
      description: 获取指定群组的所有公告
      parameters:
      - description: 群组ID
        in: path
        name: group_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: 获取群组公告列表
      tags:
      - Group
  /group/{group_id}/dissolve:
    post:
      consumes:
      - application/json
      description: 解散指定的群组，群主或管理员可以执行此操作
      parameters:
      - description: 群组ID
        in: path
        name: group_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 群组解散成功"  // 群组解散成功
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: 请求参数错误"  // 参数错误
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: 内部服务器错误"  // 服务器错误
          schema:
            $ref: '#/definitions/model.Response'
      summary: 解散群组
      tags:
      - Group
  /group/{group_id}/kick_member:
    post:
      consumes:
      - application/json
      description: 从群组中踢出指定成员
      parameters:
      - description: 群组ID
        in: path
        name: group_id
        required: true
        type: integer
      - description: 踢出成员请求数据
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.KickMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成员成功被踢出"  // 成员踢出成功
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: 请求参数错误"  // 参数错误
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: 内部服务器错误"  // 服务器错误
          schema:
            $ref: '#/definitions/model.Response'
      summary: 踢出群成员
      tags:
      - Group
  /group/{group_id}/mute:
    post:
      consumes:
      - application/json
      description: 将指定成员禁言指定时间
      parameters:
      - description: 群组ID
        in: path
        name: group_id
        required: true
        type: integer
      - description: 禁言请求数据
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.GroupMemberMuteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成员禁言成功"  // 禁言成功
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: 请求参数错误"  // 参数错误
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: 内部服务器错误"  // 服务器错误
          schema:
            $ref: '#/definitions/model.Response'
      summary: 禁言群成员
      tags:
      - Group
  /group/{group_id}/set_admin:
    post:
      consumes:
      - application/json
      description: 设置指定成员为群组管理员
      parameters:
      - description: 群组ID
        in: path
        name: group_id
        required: true
        type: integer
      - description: 设置管理员请求数据
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.SetAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 管理员设置成功"  // 管理员设置成功
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: 请求参数错误"  // 参数错误
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: 内部服务器错误"  // 服务器错误
          schema:
            $ref: '#/definitions/model.Response'
      summary: 设置群组管理员
      tags:
      - Group
  /group/{group_id}/transfer:
    post:
      consumes:
      - application/json
      description: 将群组所有权转让给其他成员
      parameters:
      - description: 群组ID
        in: path
        name: group_id
        required: true
        type: integer
      - description: 转让请求数据
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.GroupTransferRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 群组所有权转让成功"  // 转让成功
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: 请求参数错误"  // 参数错误
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: 内部服务器错误"  // 服务器错误
          schema:
            $ref: '#/definitions/model.Response'
      summary: 转让群组所有权
      tags:
      - Group
  /group/{group_id}/unmute:
    post:
      consumes:
      - application/json
      description: 解除指定成员的禁言
      parameters:
      - description: 群组ID
        in: path
        name: group_id
        required: true
        type: integer
      - description: 解除禁言请求数据
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.UnmuteMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成员解除禁言成功"  // 解除禁言成功
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: 请求参数错误"  // 参数错误
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: 内部服务器错误"  // 服务器错误
          schema:
            $ref: '#/definitions/model.Response'
      summary: 解除禁言
      tags:
      - Group
  /group/{group_id}/unset_admin:
    post:
      consumes:
      - application/json
      description: 取消指定成员的群组管理员身份
      parameters:
      - description: 群组ID
        in: path
        name: group_id
        required: true
        type: integer
      - description: 取消管理员请求数据
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.UnsetAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 管理员取消成功"  // 管理员取消成功
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: 请求参数错误"  // 参数错误
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: 内部服务器错误"  // 服务器错误
          schema:
            $ref: '#/definitions/model.Response'
      summary: 取消群组管理员
      tags:
      - Group
  /group/create:
    post:
      consumes:
      - application/json
      description: 创建一个新的群组，并将指定的成员添加到群组中
      parameters:
      - description: 创建群组请求参数
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/model.GroupCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 群组创建成功
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: 内部服务器错误
          schema:
            $ref: '#/definitions/model.Response'
      summary: 创建群组
      tags:
      - Group
  /group/join:
    get:
      consumes:
      - application/json
      description: 加入群组
      parameters:
      - description: 群组ID
        in: query
        name: group_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: 加入群组
      tags:
      - Group
  /group/member:
    get:
      consumes:
      - application/json
      description: 获取指定群组的所有成员
      parameters:
      - description: 群组ID
        in: query
        name: group_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 群组成员列表"  // 返回群组成员列表
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.MemberVo'
                  type: array
              type: object
        "400":
          description: 请求参数错误"  // 参数错误
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: 内部服务器错误"  // 服务器错误
          schema:
            $ref: '#/definitions/model.Response'
      summary: 获取群组成员
      tags:
      - Group
  /group/mute:
    post:
      consumes:
      - application/json
      description: 将群组成员禁言指定时间
      parameters:
      - description: 群组禁言请求数据
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.GroupMuteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 群组禁言成功"  // 群组禁言成功
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: 请求参数错误"  // 参数错误
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: 内部服务器错误"  // 服务器错误
          schema:
            $ref: '#/definitions/model.Response'
      summary: 群组禁言
      tags:
      - Group
  /group/quit:
    get:
      consumes:
      - application/json
      description: 退出群组
      parameters:
      - description: 群组ID
        in: query
        name: group_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: 退出群组
      tags:
      - Group
  /group/search:
    post:
      consumes:
      - application/json
      description: 搜索群组并返回搜索结果
      parameters:
      - description: 搜索群组请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.GroupSearchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 群组搜索结果"  // 返回搜索到的群组列表
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.GroupVo'
                  type: array
              type: object
        "400":
          description: 请求参数错误"  // 参数错误
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: 内部服务器错误"  // 服务器错误
          schema:
            $ref: '#/definitions/model.Response'
      summary: 搜索群组
      tags:
      - Group
  /group/update:
    post:
      consumes:
      - application/json
      description: 更新群组的名称、描述等信息
      parameters:
      - description: 更新群组信息
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/model.GroupUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 群组更新成功"  // 更新成功响应
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: 请求参数错误"  // 参数错误
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: 内部服务器错误"  // 服务器错误
          schema:
            $ref: '#/definitions/model.Response'
      summary: 更新群组信息
      tags:
      - Group
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"